+++ sm/namespaces/dbpedia.py
@@ -17,6 +17,7 @@ class DBpediaNamespace(DefaultKnowledgeGraphNamespace):
     main_namespaces: list[str] = [
         "http://dbpedia.org/ontology/",
         "http://dbpedia.org/resource/",
+        "http://dbpedia.org/property/",
     ]
 
     @classmethod
@@ -25,6 +26,13 @@ class DBpediaNamespace(DefaultKnowledgeGraphNamespace):
             {
                 "rdf": "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
                 "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
+                "owl": "http://www.w3.org/2002/07/owl#",
                 "dbo": "http://dbpedia.org/ontology/",
+                "dbr": "http://dbpedia.org/resource/",
+                "dbp": "http://dbpedia.org/property/",
+                "dc": "http://purl.org/dc/elements/1.1/",
+                "foaf": "http://xmlns.com/foaf/0.1/",
+                "xsd": "http://www.w3.org/2001/XMLSchema#",
+                "dul": "http://www.ontologydesignpatterns.org/ont/dul/DUL.owl#",
             }
         )