+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add `__str__` to `EntityIdWithScore` class
+- Add `assert_not_empty` function to check if a list is not empty
 
 ### Fixed