+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - Add missing prefixes to DBpediaNamespace and main URIs
+- Fix division by zero error in `percentage` function
 
 ## [6.10.1] - 2024-03-12